Nuevas librerías completas y probadas

+++ SmartChef/include/ultrasonic.h
@@ -1,23 +1,9 @@
 #ifndef ULTRASONIC_H
 #define ULTRASONIC_H
 
-#include "pico/stdlib.h"
+#include "pico/stdlib.h"  // Para tipo uint
 
-/**
- * @brief Inicializa los pines para el sensor ultrasónico.
- * 
- * @param trig_pin Pin de Trigger.
- * @param echo_pin Pin de Echo.
- */
-void ultrasonic_init(uint8_t trig_pin, uint8_t echo_pin);
-
-/**
- * @brief Mide la distancia en centímetros.
- * 
- * @param trig_pin Pin de Trigger.
- * @param echo_pin Pin de Echo.
- * @return La distancia medida en cm.
- */
-float ultrasonic_get_distance_cm(uint8_t trig_pin, uint8_t echo_pin);
+void ultrasonic_init(uint trig_pin, uint echo_pin);
+float ultrasonic_get_distance_cm(uint trig_pin, uint echo_pin);
 
 #endif // ULTRASONIC_H
